# gen/http/openapi3.json
{"openapi":"3.0.3","info":{"title":"Rotabot - Making rotas dead simple","description":"A service for working with rotas across multiples tools i.e Slack, Teams, etc","version":"1.0"},"servers":[{"url":"http://localhost:8080/","description":"Backend for the rotabot application."}],"paths":{"/slack/commands":{"post":{"tags":["Slack"],"summary":"Commands Slack","operationId":"Slack#Commands","parameters":[{"name":"X-Slack-Signature","in":"header","allowEmptyValue":true,"required":true,"schema":{"type":"string","example":"Omnis doloremque error natus fuga distinctio saepe."},"example":"Deserunt sint."},{"name":"X-Slack-Request-Timestamp","in":"header","allowEmptyValue":true,"required":true,"schema":{"type":"integer","example":8675811550449929158,"format":"int64"},"example":6030754581423389613}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CommandsRequestBody"},"example":{"api_app_id":"Qui sequi quia consequatur quam.","channel_id":"Dolore similique.","channel_name":"Fugit sed quam pariatur voluptatibus natus consequatur.","command":"Eos voluptas.","enterprise_id":"Maxime autem eius architecto numquam.","enterprise_name":"Aliquam blanditiis ratione repellat.","is_enterprise_install":false,"response_url":"Quas labore temporibus rerum.","team_domain":"Omnis quia cumque numquam delectus delectus.","team_id":"Cupiditate perferendis vel unde labore.","text":"Aspernatur atque explicabo eius sit quo.","token":"Velit perspiciatis.","trigger_id":"Eius dolorem.","user_id":"Sed eligendi repellendus accusamus.","user_name":"Voluptatem iure omnis."}}}},"responses":{"200":{"description":"OK response."}}}},"/slack/events":{"post":{"tags":["Slack"],"summary":"Events Slack","operationId":"Slack#Events","parameters":[{"name":"X-Slack-Signature","in":"header","allowEmptyValue":true,"required":true,"schema":{"type":"string","example":"Voluptate laborum cumque."},"example":"Delectus et."},{"name":"X-Slack-Request-Timestamp","in":"header","allowEmptyValue":true,"required":true,"schema":{"type":"integer","example":531363505557221990,"format":"int64"},"example":8131406873964025444}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventsRequestBody"},"example":{"api_app_id":"Neque blanditiis eum.","challenge":"Voluptatem est iste quam eaque.","event":{"type":"Commodi molestiae similique dignissimos quia quas."},"team_id":"Beatae exercitationem quo.","token":"Natus quo nostrum eaque eum ipsam voluptatum.","type":"Numquam qui facere hic est ea."}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventResponse"},"example":{"challenge":"randomstring"}}}}}}},"/slack/message_actions":{"post":{"tags":["Slack"],"summary":"MessageActions Slack","operationId":"Slack#MessageActions","parameters":[{"name":"X-Slack-Signature","in":"header","allowEmptyValue":true,"required":true,"schema":{"type":"string","example":"Quaerat impedit."},"example":"Voluptatem aut."},{"name":"X-Slack-Request-Timestamp","in":"header","allowEmptyValue":true,"required":true,"schema":{"type":"integer","example":2434341126659984107,"format":"int64"},"example":4179500075779313235}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageActionsRequestBody"},"example":{"payload":"TGFib3J1bSBjb25zZXF1YXR1ciBkb2xvcnVtLg=="}}}},"responses":{"200":{"description":"OK response.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActionResponse"},"example":{"errors":{"foo":"bar"},"response_action":"errors","view":"Pariatur aliquid fuga necessitatibus adipisci iusto."}}}}}}}},"components":{"schemas":{"ActionResponse":{"type":"object","properties":{"errors":{"type":"object","example":{"foo":"bar"},"additionalProperties":{"type":"string","example":"Et voluptatem."}},"response_action":{"type":"string","example":"errors"},"view":{"type":"string","example":"Esse praesentium et totam assumenda.","format":"binary"}},"example":{"errors":{"foo":"bar"},"response_action":"errors","view":"Architecto adipisci."}},"CommandsRequestBody":{"type":"object","properties":{"api_app_id":{"type":"string","example":"Ut dolor placeat at autem labore."},"channel_id":{"type":"string","example":"Minima mollitia nemo voluptates sunt."},"channel_name":{"type":"string","example":"Et sint mollitia tempora qui in est."},"command":{"type":"string","example":"Aut est architecto est."},"enterprise_id":{"type":"string","example":"Fugiat voluptatem vitae tenetur nobis minima dolorem."},"enterprise_name":{"type":"string","example":"Ut atque rerum et."},"is_enterprise_install":{"type":"boolean","example":true},"response_url":{"type":"string","example":"Repellendus eveniet."},"team_domain":{"type":"string","example":"Pariatur et corrupti."},"team_id":{"type":"string","example":"Eum saepe distinctio."},"text":{"type":"string","example":"Aut asperiores et quia."},"token":{"type":"string","example":"Quaerat qui quod vitae impedit."},"trigger_id":{"type":"string","example":"Deleniti quo veritatis ullam vel et molestiae."},"user_id":{"type":"string","example":"Atque adipisci expedita eveniet voluptas."},"user_name":{"type":"string","example":"Voluptate debitis molestiae ullam dolorem."}},"example":{"api_app_id":"Tempore quas doloribus quo aut.","channel_id":"Adipisci laborum beatae magni dolor et.","channel_name":"Explicabo et voluptatem.","command":"Perspiciatis tempora ut ut natus impedit.","enterprise_id":"Reprehenderit repellat quibusdam assumenda.","enterprise_name":"Iste quia.","is_enterprise_install":true,"response_url":"Officiis incidunt est sit.","team_domain":"Autem quas.","team_id":"Assumenda sed aut sunt consequuntur quidem.","text":"Neque earum.","token":"Nemo sit deserunt aspernatur.","trigger_id":"Suscipit magnam ut.","user_id":"Doloremque ut.","user_name":"Recusandae eos aperiam dolorem dolorem et."},"required":["token","command","trigger_id","user_id","team_id","channel_id"]},"EventResponse":{"type":"object","properties":{"challenge":{"type":"string","example":"randomstring"}},"example":{"challenge":"randomstring"}},"EventsRequestBody":{"type":"object","properties":{"api_app_id":{"type":"string","example":"Veniam sunt."},"challenge":{"type":"string","example":"Alias quisquam numquam."},"event":{"type":"object","properties":{"type":{"type":"string","example":"Eaque aut corrupti quis quo illo."}},"description":"The actual event information","example":{"type":"Est voluptatem a alias."}},"team_id":{"type":"string","example":"Et maiores voluptas inventore officia ex."},"token":{"type":"string","example":"Provident dolorem doloremque excepturi sit."},"type":{"type":"string","example":"Quo id ab expedita quis."}},"example":{"api_app_id":"Pariatur et sed aliquam iste odio.","challenge":"Exercitationem rerum quod earum minus.","event":{"type":"Animi reprehenderit sequi eos earum."},"team_id":"Non repudiandae blanditiis amet.","token":"Dolores et velit totam blanditiis quidem.","type":"Et odio perferendis."},"required":["token","team_id","type","api_app_id"]},"MessageActionsRequestBody":{"type":"object","properties":{"payload":{"type":"string","example":"SXN0ZSBub24gZWEgdmVyaXRhdGlzIHZvbHVwdGF0ZS4=","format":"binary"}},"example":{"payload":"QWNjdXNhbXVzIHZvbHVwdGFzIHJlY3VzYW5kYWUgYWxpYXMgZXQgY29uc2VxdWF0dXIgYXV0Lg=="},"required":["payload"]}}},"tags":[{"name":"Slack","description":"Slack api for interacting with slack commands, actions, events etc."}]}