# gen/http/slack/server/encode_decode.go
// Code generated by goa v3.12.3, DO NOT EDIT.
//
// Slack HTTP server encoders and decoders
//
// Command:
// $ goa gen github.com/rotabot-io/rotabot/design

package server

import (
	"context"
	"io"
	"net/http"
	"strconv"

	slack "github.com/rotabot-io/rotabot/gen/slack"
	goahttp "goa.design/goa/v3/http"
	goa "goa.design/goa/v3/pkg"
)

// EncodeCommandsResponse returns an encoder for responses returned by the
// Slack Commands endpoint.
func EncodeCommandsResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		w.WriteHeader(http.StatusOK)
		return nil
	}
}

// DecodeCommandsRequest returns a decoder for requests sent to the Slack
// Commands endpoint.
func DecodeCommandsRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			body CommandsRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateCommandsRequestBody(&body)
		if err != nil {
			return nil, err
		}

		var (
			signature string
			timestamp int64
		)
		signature = r.Header.Get("X-Slack-Signature")
		if signature == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("signature", "header"))
		}
		{
			timestampRaw := r.Header.Get("X-Slack-Request-Timestamp")
			if timestampRaw == "" {
				err = goa.MergeErrors(err, goa.MissingFieldError("timestamp", "header"))
			}
			v, err2 := strconv.ParseInt(timestampRaw, 10, 64)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("timestamp", timestampRaw, "integer"))
			}
			timestamp = v
		}
		if err != nil {
			return nil, err
		}
		payload := NewCommandsCommand(&body, signature, timestamp)

		return payload, nil
	}
}

// EncodeEventsResponse returns an encoder for responses returned by the Slack
// Events endpoint.
func EncodeEventsResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*slack.EventResponse)
		enc := encoder(ctx, w)
		body := NewEventsResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeEventsRequest returns a decoder for requests sent to the Slack Events
// endpoint.
func DecodeEventsRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			body EventsRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateEventsRequestBody(&body)
		if err != nil {
			return nil, err
		}

		var (
			signature string
			timestamp int64
		)
		signature = r.Header.Get("X-Slack-Signature")
		if signature == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("signature", "header"))
		}
		{
			timestampRaw := r.Header.Get("X-Slack-Request-Timestamp")
			if timestampRaw == "" {
				err = goa.MergeErrors(err, goa.MissingFieldError("timestamp", "header"))
			}
			v, err2 := strconv.ParseInt(timestampRaw, 10, 64)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("timestamp", timestampRaw, "integer"))
			}
			timestamp = v
		}
		if err != nil {
			return nil, err
		}
		payload := NewEventsEvent(&body, signature, timestamp)

		return payload, nil
	}
}

// EncodeMessageActionsResponse returns an encoder for responses returned by
// the Slack MessageActions endpoint.
func EncodeMessageActionsResponse(encoder func(context.Context, http.ResponseWriter) goahttp.Encoder) func(context.Context, http.ResponseWriter, any) error {
	return func(ctx context.Context, w http.ResponseWriter, v any) error {
		res, _ := v.(*slack.ActionResponse)
		enc := encoder(ctx, w)
		body := NewMessageActionsResponseBody(res)
		w.WriteHeader(http.StatusOK)
		return enc.Encode(body)
	}
}

// DecodeMessageActionsRequest returns a decoder for requests sent to the Slack
// MessageActions endpoint.
func DecodeMessageActionsRequest(mux goahttp.Muxer, decoder func(*http.Request) goahttp.Decoder) func(*http.Request) (any, error) {
	return func(r *http.Request) (any, error) {
		var (
			body MessageActionsRequestBody
			err  error
		)
		err = decoder(r).Decode(&body)
		if err != nil {
			if err == io.EOF {
				return nil, goa.MissingPayloadError()
			}
			return nil, goa.DecodePayloadError(err.Error())
		}
		err = ValidateMessageActionsRequestBody(&body)
		if err != nil {
			return nil, err
		}

		var (
			signature string
			timestamp int64
		)
		signature = r.Header.Get("X-Slack-Signature")
		if signature == "" {
			err = goa.MergeErrors(err, goa.MissingFieldError("signature", "header"))
		}
		{
			timestampRaw := r.Header.Get("X-Slack-Request-Timestamp")
			if timestampRaw == "" {
				err = goa.MergeErrors(err, goa.MissingFieldError("timestamp", "header"))
			}
			v, err2 := strconv.ParseInt(timestampRaw, 10, 64)
			if err2 != nil {
				err = goa.MergeErrors(err, goa.InvalidFieldTypeError("timestamp", timestampRaw, "integer"))
			}
			timestamp = v
		}
		if err != nil {
			return nil, err
		}
		payload := NewMessageActionsAction(&body, signature, timestamp)

		return payload, nil
	}
}